requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
Flask-SocketIO==5.3.6
gunicorn==23.0.0
Pillow==10.4.0
python-dotenv==1.0.1
eventlet==0.33.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.18.1+cpu
